+++ src/components/SQFormDialogStepper/SQFormDialogStepper.js
@@ -7,6 +7,7 @@ import {
   DialogTitle,
   Divider,
   Grid,
+  makeStyles,
   Slide,
   Step,
   StepButton,
@@ -17,7 +18,6 @@ import ArrowLeftIcon from '@material-ui/icons/KeyboardArrowLeft';
 import ArrowRightIcon from '@material-ui/icons/KeyboardArrowRight';
 import {Form, Formik, useFormikContext} from 'formik';
 import {IconButton, RoundedButton} from 'scplus-shared-components';
-import './SQFormDialogStepper.css';
 
 export function SQFormDialogStep({children}) {
   return <>{children}</>;
@@ -25,6 +25,34 @@ export function SQFormDialogStep({children}) {
 
 const Transition = React.forwardRef((props, ref) => {
   return <Slide direction="down" ref={ref} {...props} />;
+});
+
+const useStyles = makeStyles({
+  root: {
+    padding: 20,
+    '& svg': {
+      fontSize: 30
+    },
+    '& span': {
+      fontSize: 15,
+      whiteSpace: 'nowrap'
+    }
+  }
+});
+
+const useActionsStyles = makeStyles({
+  root: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    flex: '1 1 100%',
+    padding: '16px 24px'
+  }
+});
+
+const useStepperStyles = makeStyles({
+  root: {
+    justifyContent: 'center'
+  }
 });
 
 export function SQFormDialogStepper({
@@ -49,21 +77,22 @@ export function SQFormDialogStepper({
   const currentChild = steps[activeStep];
   const [completed, setCompleted] = React.useState({});
 
-  const totalSteps = React.useMemo(() => {
-    return steps.length;
-  }, [steps]);
+  const classes = useStyles();
+  const actionsClasses = useActionsStyles();
+  const stepperClasses = useStepperStyles();
 
   const isLastStep = React.useMemo(() => {
-    return activeStep === totalSteps - 1;
-  }, [activeStep, totalSteps]);
+    return activeStep === steps.length - 1;
+  }, [activeStep, steps]);
 
   // Our last step doesn't get marked complete
-  const allStepsCompleted = () =>
-    Object.keys(completed).length === totalSteps - 1;
+  const isAllStepsCompleted = () => {
+    return Object.keys(completed).length === steps.length - 1;
+  };
 
   const handleNext = () => {
     const newActiveStep =
-      isLastStep && !allStepsCompleted
+      isLastStep && !isAllStepsCompleted
         ? // It's the last step, but not all steps have been completed,
           // find the first step that has been completed
           steps.findIndex((step, i) => !(i in completed))
@@ -105,11 +134,14 @@ export function SQFormDialogStepper({
     const {errors, values} = useFormikContext();
 
     const isButtonDisabled = React.useMemo(() => {
+      if (!currentChild.props.validationSchema) {
+        return false;
+      }
       const formValues = Object.values(values).filter(val => val);
       if (!formValues.length || isLastStep) return true;
 
       if (
-        currentChild.props.validationSchema._nodes.some(step =>
+        Object.keys(currentChild.props.validationSchema.fields).some(step =>
           Object.keys(errors).includes(step)
         )
       ) {
@@ -121,6 +153,8 @@ export function SQFormDialogStepper({
 
     return (
       <IconButton
+        height="80px"
+        width="80px"
         title="Next Step"
         IconComponent={ArrowRightIcon}
         isDisabled={isButtonDisabled}
@@ -134,14 +168,14 @@ export function SQFormDialogStepper({
     const {errors, values} = useFormikContext();
 
     const isButtonDisabled = React.useMemo(() => {
-      const currentStepKeys = currentChild.props.validationSchema._nodes;
-      const currentStepValues = Object.keys(values)
-        .filter(key => currentStepKeys.includes(key))
-        .reduce((obj, key) => {
-          obj[key] = values[key];
-          return obj;
-        }, {});
-      const formValues = Object.values(currentStepValues).filter(val => val);
+      if (!currentChild.props.validationSchema) {
+        return false;
+      }
+      const currentStepKeys = Object.keys(
+        currentChild.props.validationSchema.fields
+      );
+      const formValues = Object.values(values).filter(val => val);
+
       if (
         !formValues.length ||
         currentStepKeys.some(step => Object.keys(errors).includes(step))
@@ -149,10 +183,10 @@ export function SQFormDialogStepper({
         return true;
       }
 
-      if (allStepsCompleted() && isLastStep) {
+      if (isLastStep && isAllStepsCompleted()) {
         return false;
       }
-      return true;
+      return false;
     }, [errors, values]);
     return (
       <RoundedButton
@@ -160,7 +194,7 @@ export function SQFormDialogStepper({
         isDisabled={isButtonDisabled}
         title={cancelButtonText}
       >
-        Submit
+        {isLastStep ? 'Submit' : 'Next'}
       </RoundedButton>
     );
   }
@@ -184,35 +218,41 @@ export function SQFormDialogStepper({
             <DialogTitle disableTypography={true}>
               <Typography variant="h4">{title}</Typography>
             </DialogTitle>
-            <DialogContent
-              dividers={true}
-              style={{...contentStyle, paddingTop: '.75rem'}}
-            >
-              <div className="SQFormDialogStepper__stepContainer">
+            <Divider />
+            {steps.length > 1 && (
+              <Grid container classes={stepperClasses}>
                 <IconButton
+                  height="80px"
+                  width="80px"
                   title="Previous Step"
                   IconComponent={ArrowLeftIcon}
                   isDisabled={activeStep === 0}
                   isIconTeal={true}
                   onClick={handleBack}
                 />
-                <div className="SQFormDialogStepper__stepper">
-                  <Stepper nonLinear activeStep={activeStep}>
-                    {steps.map((child, index) => (
-                      <Step key={child.props.label}>
-                        <StepButton
-                          onClick={handleStep(index)}
-                          completed={completed[index]}
-                        >
-                          {child?.props.label}
-                        </StepButton>
-                      </Step>
-                    ))}
-                  </Stepper>
-                </div>
+                <Stepper nonLinear activeStep={activeStep} classes={classes}>
+                  {steps.map((child, index) => (
+                    <Step key={child.props.label}>
+                      <StepButton
+                        onClick={handleStep(index)}
+                        completed={completed[index]}
+                      >
+                        {child?.props.label}
+                      </StepButton>
+                    </Step>
+                  ))}
+                </Stepper>
                 <NextButton />
-              </div>
-              <Divider className="SQFormDialogStepper__divider" />
+              </Grid>
+            )}
+            <DialogContent
+              dividers
+              style={{
+                paddingTop: '40px',
+                paddingBottom: '40px',
+                ...contentStyle
+              }}
+            >
               <Grid
                 {...muiGridProps}
                 container
@@ -222,7 +262,7 @@ export function SQFormDialogStepper({
                 {currentChild}
               </Grid>
             </DialogContent>
-            <DialogActions className="SQFormDialogStepper__actions">
+            <DialogActions classes={actionsClasses}>
               <RoundedButton
                 title={cancelButtonText}
                 onClick={onClose}
@@ -243,8 +283,8 @@ export function SQFormDialogStepper({
 SQFormDialogStepper.propTypes = {
   /** The secondary button text (Button located on left side of Dialog) */
   cancelButtonText: PropTypes.string,
-  /** The content to be rendered in the dialog body */
-  children: PropTypes.node.isRequired,
+  /** The content to be rendered in the dialog body.  Will be an array of React elements. */
+  children: PropTypes.array.isRequired,
   /** If true, clicking the backdrop will not fire the onClose callback. */
   disableBackdropClick: PropTypes.bool,
   /** Sets the dialog to the maxWidth. */